requirements.txt
Pillow==9.2.0
Pillow==9.4.0
Pillow==10.0.1
Pillow==10.3.0
streamlit==1.35.0
